[INT-1265] Use server-side state for branch_uri  (#10018)
Unfortunately this issue revealed that there are a few more places
outside of our control that Jupyterlab interacts with the Contents api.
I tried to find a way to maybe override the headers or cookies sent for
these requests, but there was no relevant api.

As a result I refactored the branch_uri state to be in pfs_manager on
the server side. There is a new `/mount` api that accepts a branch_uri
parameter to set the mounted_branch. Calling that api with branch_uri=''
clears any mounted branch.

Tests have been updated. One minor thing to call attention to is that I
removed the delete all call in `test_integrations` in favor of deleting
each repo in the tests. That way we do not have to clear all our test
data out to run the tests.